Change nvrhi ClearDepthStencilAttachment() to use uint8_t for stencil type consistency
(cherry picked from commit 9eb2a3a1ee53c17c7b00669a88bae7c27a46178a)

+++ include/nvrhi/utils.h
@@ -60,7 +60,7 @@ namespace nvrhi::utils
         ICommandList* commandList,
         IFramebuffer* framebuffer,
         float depth,
-        uint32_t stencil
+        uint8_t stencil
     );
 
     NVRHI_API void BuildBottomLevelAccelStruct(
